refactor: main modularization
Main modularization for key and keyless event handling for better
understanding and readability of the code.
