+++ src/app/screens/latest.rs
@@ -97,7 +97,7 @@ impl LatestPatchsetsState {
             .get_patch_feed_page(self.page_size, self.page_number)
     }
 
-    pub fn get_number_of_processed_patchsets(&self) -> usize {
+    pub fn processed_patchsets_count(&self) -> usize {
         self.lore_session.representative_patches_ids().len()
     }
 }
